move ArifactDeckDecoder class to sperated file and add the encoder
